backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
aioboto3==13.2.0
pydantic-settings==2.7.0
